pyproject.toml
[tool.poetry]
name = "passwordgenerator"
version = "0.1.0"
description = ""
authors = ["Raul Catalinas <105791463+RaulCatalinas@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
flet = "^0.5.0"
python-dotenv = "^1.0.0"
pandas = "^1.5.3"
openpyxl = "^3.1.2"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pyinstaller = "^5.9.0"
pylint = "^2.17.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
